refactor(T1427): split 4 oversized pages + fix type casts + perf optimization
Component splits (4 pages → 20 sub-components):
- kpi/page.tsx: 1532 → 320 lines (6 components extracted)
- admin/page.tsx: 1401 → 176 lines (4 components extracted)
- reports/page.tsx: 1267 → 133 lines (3 components extracted)
- kanban/page.tsx: 1056 → 614 lines (3 components extracted)

Type safety fixes:
- time-entries/templates: replace 4x `as unknown as { userId }` with proper select
- Remove 5 unused Zod schemas (goal-validators, knowledge-validators)

Performance:
- monthly-summary: O(N×M) filter → Map lookup + single-pass approval count

Net: -4033 lines from main pages, +20 focused components

Closes #1427
