+++ .env.recovered
@@ -1,0 +1,75 @@
+AUTH_TRUST_HOST=true
+NEXTAUTH_SECRET=7bfcab377c76469e60893adfa13bd9de934ac8ab2f6993796856a1852583f017
+CRON_SECRET=ffb7d4978e274afd622b803f7b62bfd0595a998c0ba72f5818ab77054a4ad901
+DATABASE_URL=postgresql://titan:QNU9juSCkfnTXCIFqoeu@postgres:5432/titan?connection_limit=10
+REDIS_URL=redis://:n6iETYQblJTH8oGNCCWx@redis:6379
+AUTH_SECRET=7bfcab377c76469e60893adfa13bd9de934ac8ab2f6993796856a1852583f017
+NEXTAUTH_URL=https://mac-mini.tailde842d.ts.net
+AUTH_URL=https://mac-mini.tailde842d.ts.net
+YARN_VERSION=1.22.22
+NEXT_TELEMETRY_DISABLED=1
+PORT=3100
+HOSTNAME=0.0.0.0
+
+POSTGRES_USER=
+POSTGRES_PASSWORD=
+POSTGRES_DB=
+POSTGRES_PORT=
+POSTGRES_APP_USER=
+POSTGRES_APP_PASSWORD=
+REDIS_PASSWORD=
+REDIS_PORT=
+TITAN_URL=
+NGINX_HTTP_PORT=
+NGINX_HTTPS_PORT=
+MINIO_ROOT_USER=
+MINIO_ROOT_PASSWORD=
+MINIO_API_PORT=
+MINIO_CONSOLE_PORT=
+MINIO_REGION=
+OUTLINE_SECRET_KEY=
+OUTLINE_UTILS_SECRET=
+OUTLINE_URL=
+OUTLINE_PORT=
+OUTLINE_S3_BUCKET=
+OUTLINE_LOG_LEVEL=
+NEXT_PUBLIC_OUTLINE_URL=
+OUTLINE_INTERNAL_URL=
+OIDC_CLIENT_ID=
+OIDC_CLIENT_SECRET=
+OIDC_AUTH_URI=
+OIDC_TOKEN_URI=
+OIDC_USERINFO_URI=
+OIDC_USERNAME_CLAIM=
+SMTP_HOST=
+SMTP_PORT=
+SMTP_USERNAME=
+SMTP_PASSWORD=
+SMTP_FROM_EMAIL=
+SMTP_SECURE=
+SMTP_FROM=
+SMTP_USER=
+ALERT_EMAIL_TO=
+ALERT_WEBHOOK_URL=
+HOMEPAGE_PUID=
+HOMEPAGE_PGID=
+HOMEPAGE_ALLOWED_HOSTS=
+HOMEPAGE_PORT=
+HOMEPAGE_VAR_OUTLINE_URL=
+HOMEPAGE_VAR_MINIO_CONSOLE_URL=
+HOMEPAGE_VAR_MINIO_ROOT_USER=
+REPLICATION_USER=
+REPLICATION_PASSWORD=
+REPLICATION_SLOT=
+FAILOVER_CHECK_INTERVAL=
+FAILOVER_FAILURE_THRESHOLD=
+FAILOVER_CHECK_TIMEOUT=
+LDAP_URL=
+LDAP_BIND_DN=
+LDAP_BIND_PASSWORD=
+LDAP_BASE_DN=
+MIN_MOBILE_VERSION=
+EXPO_ACCESS_TOKEN=
+MONITORING_WEBHOOK_KEY=
+SLOW_QUERY_THRESHOLD_MS=
+SLOW_REQUEST_THRESHOLD_MS=
