+++ app/(app)/kanban/page.tsx
@@ -3,27 +3,21 @@
 /**
  * Kanban Page — Issue #803 (K-1)
  *
- * Features:
- * - Drag-and-drop between columns with optimistic update + rollback
- * - Same-column reordering (position-based)
- * - Custom column names (persisted in localStorage)
- * - Keyboard navigation (Tab, Enter, Arrow keys)
- * - Visual drag feedback (drop indicators, column highlight)
- * - Activity log integration via API
- * - Multi-select mode with batch operations (Issue #1023)
+ * Orchestrator: holds all state + business logic; delegates rendering to
+ *   - KanbanFilters  (filter bar)
+ *   - KanbanBoard    (5-column board)
+ *   - BulkActionBar / BulkOperationToast / TemplateImportDialog (task-quick-edit)
+ *   - TaskDetailModal
  */
 
 import { useState, useEffect, useCallback, useMemo, useRef } from "react";
 import { useSearchParams } from "next/navigation";
 import {
   Plus,
-  Kanban,
   Loader2,
-  CheckSquare,
-  X,
   MousePointerSquareDashed,
-  Check,
   GripVertical,
+  X,
   FileInput,
 } from "lucide-react";
 import { toast } from "sonner";
@@ -31,70 +25,23 @@ import { cn } from "@/lib/utils";
 import { extractItems } from "@/lib/api-client";
 import { type TaskCardData } from "@/app/components/task-card";
 import {
-  TaskFilters,
   type TaskFilters as FiltersType,
-  emptyFilters,
-  hasActiveFilters,
   parseFiltersFromUrl,
 } from "@/app/components/task-filters";
 import { TaskDetailModal } from "@/app/components/task-detail-modal";
-import { PageLoading, PageError, PageEmpty } from "@/app/components/page-states";
-import { KanbanColumn } from "@/app/components/kanban-column";
 import { calculatePosition } from "@/lib/utils/optimistic-update";
+import { KanbanFilters } from "@/app/components/kanban/kanban-filters";
+import { KanbanBoard, type TaskStatus, COLUMN_ORDER } from "@/app/components/kanban/kanban-board";
+import {
+  BulkActionBar,
+  BulkOperationToast,
+  TemplateImportDialog,
+  type BulkToast,
+} from "@/app/components/kanban/task-quick-edit";
 
-type TaskStatus = "BACKLOG" | "TODO" | "IN_PROGRESS" | "REVIEW" | "DONE";
-
-const COLUMN_ORDER: TaskStatus[] = [
-  "BACKLOG",
-  "TODO",
-  "IN_PROGRESS",
-  "REVIEW",
-  "DONE",
-];
-
-const DEFAULT_COLUMNS: { status: TaskStatus; label: string; color: string }[] =
-  [
-    { status: "BACKLOG", label: "待辦清單", color: "text-muted-foreground" },
-    { status: "TODO", label: "待處理", color: "text-blue-400" },
-    { status: "IN_PROGRESS", label: "進行中", color: "text-yellow-400" },
-    { status: "REVIEW", label: "審核中", color: "text-purple-400" },
-    { status: "DONE", label: "已完成", color: "text-emerald-400" },
-  ];
-
-const columnBorder: Record<TaskStatus, string> = {
-  BACKLOG: "border-border",
-  TODO: "border-blue-500/20",
-  IN_PROGRESS: "border-yellow-500/20",
-  REVIEW: "border-purple-500/20",
-  DONE: "border-emerald-500/20",
-};
-
-const columnHeaderBg: Record<TaskStatus, string> = {
-  BACKLOG: "bg-muted/60",
-  TODO: "bg-blue-500/10",
-  IN_PROGRESS: "bg-yellow-500/10",
-  REVIEW: "bg-purple-500/10",
-  DONE: "bg-emerald-500/10",
-};
-
-const STATUS_OPTIONS: { value: TaskStatus; label: string }[] = [
-  { value: "BACKLOG", label: "待辦清單" },
-  { value: "TODO", label: "待處理" },
-  { value: "IN_PROGRESS", label: "進行中" },
-  { value: "REVIEW", label: "審核中" },
-  { value: "DONE", label: "已完成" },
-];
-
-const PRIORITY_OPTIONS = [
-  { value: "P0", label: "P0" },
-  { value: "P1", label: "P1" },
-  { value: "P2", label: "P2" },
-  { value: "P3", label: "P3" },
-];
+// ─── Local helpers ────────────────────────────────────────────────────────────
 
 const COLUMN_NAMES_KEY = "titan-kanban-column-names";
-
-const EMPTY_TASK_ARRAY: (TaskCardData & { position?: number })[] = [];
 
 function loadColumnNames(): Record<string, string> {
   if (typeof window === "undefined") return {};
@@ -115,16 +62,14 @@ function saveColumnNames(names: Record<string, string>) {
   }
 }
 
+// ─── Page ─────────────────────────────────────────────────────────────────────
+
 export default function KanbanPage() {
   const searchParams = useSearchParams();
-  const [tasks, setTasks] = useState<
-    (TaskCardData & { position?: number })[]
-  >([]);
+  const [tasks, setTasks] = useState<(TaskCardData & { position?: number })[]>([]);
   const [loading, setLoading] = useState(true);
   const [fetchError, setFetchError] = useState<string | null>(null);
-  const [allTasks, setAllTasks] = useState<
-    (TaskCardData & { position?: number })[]
-  >([]);
+  const [allTasks, setAllTasks] = useState<(TaskCardData & { position?: number })[]>([]);
   const [filters, setFilters] = useState<FiltersType>(() =>
     parseFiltersFromUrl(new URLSearchParams(searchParams.toString()))
   );
@@ -134,9 +79,7 @@ export default function KanbanPage() {
   // Issue #1231: auto-open task detail modal from URL ?task= parameter
   useEffect(() => {
     const taskParam = searchParams.get("task");
-    if (taskParam) {
-      setSelectedTaskId(taskParam);
-    }
+    if (taskParam) setSelectedTaskId(taskParam);
   }, [searchParams]);
 
   useEffect(() => {
@@ -144,6 +87,7 @@ export default function KanbanPage() {
       setShowDragHint(true);
     }
   }, []);
+
   const [dragOver, setDragOver] = useState<TaskStatus | null>(null);
   const [draggingId, setDraggingId] = useState<string | null>(null);
   const [movingTask, setMovingTask] = useState<string | null>(null);
@@ -153,10 +97,7 @@ export default function KanbanPage() {
   const [multiSelectMode, setMultiSelectMode] = useState(false);
   const [selectedIds, setSelectedIds] = useState<Set<string>>(new Set());
   const [bulkLoading, setBulkLoading] = useState(false);
-  const [bulkToast, setBulkToast] = useState<{
-    type: "success" | "error" | "partial";
-    message: string;
-  } | null>(null);
+  const [bulkToast, setBulkToast] = useState<BulkToast | null>(null);
   const [users, setUsers] = useState<{ id: string; name: string }[]>([]);
 
   // Project filter state (Issue #1176)
@@ -173,17 +114,16 @@ export default function KanbanPage() {
   const tasksRef = useRef(tasks);
   useEffect(() => { tasksRef.current = tasks; }, [tasks]);
 
-  // Ref to always access latest moveTask without stale closures in useCallback handlers
+  // Ref to always access latest moveTask without stale closures
   const moveTaskRef = useRef<(taskId: string, newStatus: TaskStatus, targetIndex?: number) => Promise<void>>(async () => {});
-
 
   // Snapshot for rollback
   const tasksSnapshot = useRef<(TaskCardData & { position?: number })[]>([]);
 
   // Load custom column names from localStorage
-  useEffect(() => {
-    setColumnNames(loadColumnNames());
-  }, []);
+  useEffect(() => { setColumnNames(loadColumnNames()); }, []);
+
+  // ── Data fetching ─────────────────────────────────────────────────────────
 
   const fetchTasks = useCallback(async () => {
     setLoading(true);
@@ -199,22 +139,13 @@ export default function KanbanPage() {
       const res = await fetch(`/api/tasks?${params}`);
       if (!res.ok) throw new Error("任務載入失敗");
       const body = await res.json();
-      const fetched = extractItems<
-        TaskCardData & {
-          position?: number;
-          tags?: string[];
-          dueDate?: string | null;
-        }
-      >(body);
+      const fetched = extractItems<TaskCardData & { position?: number; tags?: string[]; dueDate?: string | null }>(body);
       setAllTasks(fetched);
 
-      // Client-side AND filtering for tags and dueDate range
       let filtered = fetched;
       if (filters.tags.length > 0) {
         filtered = filtered.filter((t) =>
-          filters.tags.every((tag) =>
-            (t as unknown as { tags?: string[] }).tags?.includes(tag)
-          )
+          filters.tags.every((tag) => (t as unknown as { tags?: string[] }).tags?.includes(tag))
         );
       }
       if (filters.dueDateFrom) {
@@ -237,11 +168,9 @@ export default function KanbanPage() {
     }
   }, [filters, projectFilter]);
 
-  useEffect(() => {
-    fetchTasks();
-  }, [fetchTasks]);
+  useEffect(() => { fetchTasks(); }, [fetchTasks]);
 
-  // Fetch users for assignee picker
+  // Fetch users + projects for filters
   useEffect(() => {
     fetch("/api/users")
       .then((r) => r.json())
@@ -250,7 +179,6 @@ export default function KanbanPage() {
         setUsers(list.map((u) => ({ id: u.id, name: u.name })));
       })
       .catch(() => { toast.error("使用者清單載入失敗"); });
-    // Fetch project options for filter (Issue #1176)
     fetch("/api/projects?limit=100")
       .then((r) => r.json())
       .then((body) => {
@@ -261,62 +189,41 @@ export default function KanbanPage() {
       .catch(() => { toast.error("專案清單載入失敗"); });
   }, []);
 
-  // ── Column name management ──────────────────────────────────────────────
+  // ── Column name management ────────────────────────────────────────────────
 
-  const handleColumnNameChange = useCallback(
-    (status: TaskStatus, name: string) => {
-      setColumnNames((prev) => {
-        const next = { ...prev, [status]: name };
-        saveColumnNames(next);
-        return next;
-      });
-    },
-    []
-  );
+  const handleColumnNameChange = useCallback((status: TaskStatus, name: string) => {
+    setColumnNames((prev) => {
+      const next = { ...prev, [status]: name };
+      saveColumnNames(next);
+      return next;
+    });
+  }, []);
 
-  // ── Move task between columns (optimistic) ─────────────────────────────
+  // ── Move task between columns (optimistic) ────────────────────────────────
 
-  async function moveTask(
-    taskId: string,
-    newStatus: TaskStatus,
-    targetIndex?: number
-  ) {
+  async function moveTask(taskId: string, newStatus: TaskStatus, targetIndex?: number) {
     const task = tasks.find((t) => t.id === taskId);
     if (!task || task.status === newStatus) {
-      // Same column reorder
       if (task && targetIndex !== undefined) {
         await reorderInColumn(taskId, task.status as TaskStatus, targetIndex);
       }
       return;
     }
 
-    // Snapshot for rollback
     tasksSnapshot.current = [...tasks];
 
-    // Calculate position for insertion
     const columnTasks = tasks
       .filter((t) => t.status === newStatus && t.id !== taskId)
       .sort((a, b) => (a.position ?? 0) - (b.position ?? 0));
 
     const insertIdx = targetIndex ?? columnTasks.length;
-    const prevPos =
-      insertIdx > 0 ? (columnTasks[insertIdx - 1]?.position ?? 0) : null;
-    const nextPos =
-      insertIdx < columnTasks.length
-        ? (columnTasks[insertIdx]?.position ?? 0)
-        : null;
+    const prevPos = insertIdx > 0 ? (columnTasks[insertIdx - 1]?.position ?? 0) : null;
+    const nextPos = insertIdx < columnTasks.length ? (columnTasks[insertIdx]?.position ?? 0) : null;
     const newPosition = calculatePosition(prevPos, nextPos);
 
-    // Optimistic update
     setTasks((prev) =>
       prev.map((t) =>
-        t.id === taskId
-          ? {
-              ...t,
-              status: newStatus as TaskCardData["status"],
-              position: newPosition,
-            }
-          : t
+        t.id === taskId ? { ...t, status: newStatus as TaskCardData["status"], position: newPosition } : t
       )
     );
     setMovingTask(taskId);
@@ -328,17 +235,13 @@ export default function KanbanPage() {
         body: JSON.stringify({ status: newStatus }),
       });
       if (!res.ok) {
-        // Rollback
         setTasks(tasksSnapshot.current);
         toast.error("任務更新失敗");
       } else {
-        // Also update position via reorder API
         const reorderRes = await fetch("/api/tasks/reorder", {
           method: "POST",
           headers: { "Content-Type": "application/json" },
-          body: JSON.stringify({
-            items: [{ id: taskId, position: newPosition, status: newStatus }],
-          }),
+          body: JSON.stringify({ items: [{ id: taskId, position: newPosition, status: newStatus }] }),
         });
         if (reorderRes.ok) {
           toast.success("任務已更新");
@@ -354,16 +257,11 @@ export default function KanbanPage() {
     }
   }
 
-  // Keep ref in sync so useCallback handlers can call moveTask without stale closures
   moveTaskRef.current = moveTask;
 
-  // ── Reorder within same column ────────────────────────────────────────
+  // ── Reorder within same column ────────────────────────────────────────────
 
-  async function reorderInColumn(
-    taskId: string,
-    status: TaskStatus,
-    targetIndex: number
-  ) {
+  async function reorderInColumn(taskId: string, status: TaskStatus, targetIndex: number) {
     const columnTasks = tasks
       .filter((t) => t.status === status)
       .sort((a, b) => (a.position ?? 0) - (b.position ?? 0));
@@ -371,139 +269,93 @@ export default function KanbanPage() {
     const currentIndex = columnTasks.findIndex((t) => t.id === taskId);
     if (currentIndex === targetIndex || currentIndex === -1) return;
 
-    // Calculate new position
     const filtered = columnTasks.filter((t) => t.id !== taskId);
-    const prevPos =
-      targetIndex > 0 ? (filtered[targetIndex - 1]?.position ?? 0) : null;
-    const nextPos =
-      targetIndex < filtered.length
-        ? (filtered[targetIndex]?.position ?? 0)
-        : null;
+    const prevPos = targetIndex > 0 ? (filtered[targetIndex - 1]?.position ?? 0) : null;
+    const nextPos = targetIndex < filtered.length ? (filtered[targetIndex]?.position ?? 0) : null;
     const newPosition = calculatePosition(prevPos, nextPos);
 
-    // Snapshot for rollback
     tasksSnapshot.current = [...tasks];
-
-    // Optimistic update
-    setTasks((prev) =>
-      prev.map((t) => (t.id === taskId ? { ...t, position: newPosition } : t))
-    );
+    setTasks((prev) => prev.map((t) => (t.id === taskId ? { ...t, position: newPosition } : t)));
 
     try {
       await fetch("/api/tasks/reorder", {
         method: "POST",
         headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({
-          items: [{ id: taskId, position: newPosition }],
-        }),
+        body: JSON.stringify({ items: [{ id: taskId, position: newPosition }] }),
       });
     } catch {
       setTasks(tasksSnapshot.current);
     }
   }
 
-  // ── Drag & drop handlers ──────────────────────────────────────────────
+  // ── Drag & drop handlers ──────────────────────────────────────────────────
 
-  const handleDragStart = useCallback(
-    (e: React.DragEvent, taskId: string) => {
-      e.dataTransfer.setData("taskId", taskId);
-      e.dataTransfer.effectAllowed = "move";
-      setDraggingId(taskId);
-    },
-    []
-  );
+  const handleDragStart = useCallback((e: React.DragEvent, taskId: string) => {
+    e.dataTransfer.setData("taskId", taskId);
+    e.dataTransfer.effectAllowed = "move";
+    setDraggingId(taskId);
+  }, []);
 
   const handleDragEnd = useCallback(() => {
     setDraggingId(null);
     setDragOver(null);
   }, []);
 
-  const handleDragOver = useCallback(
-    (e: React.DragEvent, status: TaskStatus) => {
-      e.preventDefault();
-      e.dataTransfer.dropEffect = "move";
-      setDragOver(status);
-    },
-    []
-  );
-
-  const handleDragLeave = useCallback(() => {
-    setDragOver(null);
+  const handleDragOver = useCallback((e: React.DragEvent, status: TaskStatus) => {
+    e.preventDefault();
+    e.dataTransfer.dropEffect = "move";
+    setDragOver(status);
   }, []);
 
-  const handleDrop = useCallback(
-    (e: React.DragEvent, status: TaskStatus) => {
-      e.preventDefault();
-      const taskId = e.dataTransfer.getData("taskId");
-      if (taskId) {
-        // Calculate target index from drop position
-        const container = e.currentTarget as HTMLElement;
-        const cards = container.querySelectorAll("[data-task-id]");
-        let targetIdx = tasksRef.current.filter((t) => t.status === status).length;
-        for (let i = 0; i < cards.length; i++) {
-          const rect = cards[i].getBoundingClientRect();
-          if (e.clientY < rect.top + rect.height / 2) {
-            targetIdx = i;
-            break;
-          }
-        }
-        moveTaskRef.current(taskId, status, targetIdx);
+  const handleDragLeave = useCallback(() => { setDragOver(null); }, []);
+
+  const handleDrop = useCallback((e: React.DragEvent, status: TaskStatus) => {
+    e.preventDefault();
+    const taskId = e.dataTransfer.getData("taskId");
+    if (taskId) {
+      const container = e.currentTarget as HTMLElement;
+      const cards = container.querySelectorAll("[data-task-id]");
+      let targetIdx = tasksRef.current.filter((t) => t.status === status).length;
+      for (let i = 0; i < cards.length; i++) {
+        const rect = cards[i].getBoundingClientRect();
+        if (e.clientY < rect.top + rect.height / 2) { targetIdx = i; break; }
       }
-      setDragOver(null);
-      setDraggingId(null);
-    },
-    []
-  );
+      moveTaskRef.current(taskId, status, targetIdx);
+    }
+    setDragOver(null);
+    setDraggingId(null);
+  }, []);
 
-  // ── Keyboard navigation ───────────────────────────────────────────────
+  // ── Keyboard navigation ───────────────────────────────────────────────────
 
-  const handleKeyboardMove = useCallback(
-    (taskId: string, direction: "left" | "right") => {
-      const currentTasks = tasksRef.current;
-      const task = currentTasks.find((t) => t.id === taskId);
-      if (!task) return;
+  const handleKeyboardMove = useCallback((taskId: string, direction: "left" | "right") => {
+    const task = tasksRef.current.find((t) => t.id === taskId);
+    if (!task) return;
+    const currentIdx = COLUMN_ORDER.indexOf(task.status as TaskStatus);
+    const newIdx = direction === "left" ? currentIdx - 1 : currentIdx + 1;
+    if (newIdx < 0 || newIdx >= COLUMN_ORDER.length) return;
+    moveTask(taskId, COLUMN_ORDER[newIdx]);
+  }, []); // eslint-disable-line react-hooks/exhaustive-deps
 
-      const currentIdx = COLUMN_ORDER.indexOf(task.status as TaskStatus);
-      const newIdx = direction === "left" ? currentIdx - 1 : currentIdx + 1;
-      if (newIdx < 0 || newIdx >= COLUMN_ORDER.length) return;
+  const handleKeyboardReorder = useCallback((taskId: string, direction: "up" | "down") => {
+    const task = tasksRef.current.find((t) => t.id === taskId);
+    if (!task) return;
+    const columnTasks = tasksRef.current
+      .filter((t) => t.status === task.status)
+      .sort((a, b) => (a.position ?? 0) - (b.position ?? 0));
+    const currentIdx = columnTasks.findIndex((t) => t.id === taskId);
+    const newIdx = direction === "up" ? currentIdx - 1 : currentIdx + 1;
+    if (newIdx < 0 || newIdx >= columnTasks.length) return;
+    reorderInColumn(taskId, task.status as TaskStatus, newIdx);
+  }, []); // eslint-disable-line react-hooks/exhaustive-deps
 
-      const newStatus = COLUMN_ORDER[newIdx];
-      moveTask(taskId, newStatus);
-    },
-    [] // eslint-disable-line react-hooks/exhaustive-deps
-  );
-
-  const handleKeyboardReorder = useCallback(
-    (taskId: string, direction: "up" | "down") => {
-      const currentTasks = tasksRef.current;
-      const task = currentTasks.find((t) => t.id === taskId);
-      if (!task) return;
-
-      const columnTasks = currentTasks
-        .filter((t) => t.status === task.status)
-        .sort((a, b) => (a.position ?? 0) - (b.position ?? 0));
-
-      const currentIdx = columnTasks.findIndex((t) => t.id === taskId);
-      const newIdx = direction === "up" ? currentIdx - 1 : currentIdx + 1;
-      if (newIdx < 0 || newIdx >= columnTasks.length) return;
-
-      reorderInColumn(taskId, task.status as TaskStatus, newIdx);
-    },
-    [] // eslint-disable-line react-hooks/exhaustive-deps
-  );
-
-  // ── Create task (used by header button and empty state CTA) ──────────────
+  // ── Create task ───────────────────────────────────────────────────────────
 
   async function handleCreateTask() {
     const res = await fetch("/api/tasks", {
       method: "POST",
       headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        title: "新任務",
-        status: "BACKLOG",
-        priority: "P2",
-        category: "PLANNED",
-      }),
+      body: JSON.stringify({ title: "新任務", status: "BACKLOG", priority: "P2", category: "PLANNED" }),
     });
     if (res.ok) {
       const created = await res.json();
@@ -516,35 +368,23 @@ export default function KanbanPage() {
     }
   }
 
-  // ── Multi-select mode toggle (Issue #1023) ─────────────────────────────
+  // ── Multi-select (Issue #1023) ────────────────────────────────────────────
 
   function toggleMultiSelectMode() {
-    if (multiSelectMode) {
-      // Exiting multi-select: clear selection
-      setSelectedIds(new Set());
-    }
+    if (multiSelectMode) setSelectedIds(new Set());
     setMultiSelectMode((prev) => !prev);
   }
-
-  // ── Bulk selection helpers ────────────────────────────────────────────
 
   const toggleSelect = useCallback((taskId: string) => {
     setSelectedIds((prev) => {
       const next = new Set(prev);
-      if (next.has(taskId)) {
-        next.delete(taskId);
-      } else {
-        next.add(taskId);
-      }
+      if (next.has(taskId)) { next.delete(taskId); } else { next.add(taskId); }
       return next;
     });
   }, []);
 
-  function clearSelection() {
-    setSelectedIds(new Set());
-  }
+  function clearSelection() { setSelectedIds(new Set()); }
 
-  // Issue #1023: batch operations via Promise.allSettled with concurrency limit
   async function executeBulkAction(updates: Record<string, unknown>) {
     if (selectedIds.size === 0) return;
     setBulkLoading(true);
@@ -561,35 +401,20 @@ export default function KanbanPage() {
               method: "PATCH",
               headers: { "Content-Type": "application/json" },
               body: JSON.stringify(updates),
-            }).then((r) => {
-              if (!r.ok) throw new Error(`HTTP ${r.status}`);
-              return r;
-            })
+            }).then((r) => { if (!r.ok) throw new Error(`HTTP ${r.status}`); return r; })
           )
         );
         results.push(...chunkResults);
       }
-      const succeeded = results.filter(
-        (r) => r.status === "fulfilled"
-      ).length;
+      const succeeded = results.filter((r) => r.status === "fulfilled").length;
       const failed = results.length - succeeded;
       if (failed === 0) {
-        setBulkToast({
-          type: "success",
-          message: `${succeeded} 項已更新`,
-        });
+        setBulkToast({ type: "success", message: `${succeeded} 項已更新` });
         clearSelection();
       } else if (succeeded === 0) {
-        setBulkToast({
-          type: "error",
-          message: `${failed} 項更新失敗`,
-        });
+        setBulkToast({ type: "error", message: `${failed} 項更新失敗` });
       } else {
-        setBulkToast({
-          type: "partial",
-          message: `${succeeded} 項已更新，${failed} 項失敗`,
-        });
-        // Keep only failed task IDs selected for retry
+        setBulkToast({ type: "partial", message: `${succeeded} 項已更新，${failed} 項失敗` });
         const failedIds = ids.filter((_, idx) => results[idx].status === "rejected");
         setSelectedIds(new Set(failedIds));
       }
@@ -600,24 +425,15 @@ export default function KanbanPage() {
     }
   }
 
-  // ── Template import (Issue #1266) ────────────────────────────────────────
+  // ── Template import (Issue #1266) ─────────────────────────────────────────
 
   async function handleImportTemplate() {
     setImportError(null);
     let parsed: unknown;
-    try {
-      parsed = JSON.parse(importJson);
-    } catch {
-      setImportError("JSON 格式錯誤，請確認格式正確");
-      return;
-    }
+    try { parsed = JSON.parse(importJson); } catch { setImportError("JSON 格式錯誤，請確認格式正確"); return; }
 
-    // Accept either { tasks: [...] } or direct array
-    const tasks = Array.isArray(parsed)
-      ? parsed
-      : (parsed as { tasks?: unknown[] }).tasks;
-
-    if (!Array.isArray(tasks) || tasks.length === 0) {
+    const taskList = Array.isArray(parsed) ? parsed : (parsed as { tasks?: unknown[] }).tasks;
+    if (!Array.isArray(taskList) || taskList.length === 0) {
       setImportError("請提供至少一筆任務（tasks 陣列）");
       return;
     }
@@ -627,23 +443,17 @@ export default function KanbanPage() {
       const res = await fetch("/api/tasks/import-template", {
         method: "POST",
         headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ tasks }),
+        body: JSON.stringify({ tasks: taskList }),
       });
       const body = await res.json();
-      if (!res.ok) {
-        setImportError(body?.message ?? body?.error ?? "匯入失敗");
-        return;
-      }
+      if (!res.ok) { setImportError(body?.message ?? body?.error ?? "匯入失敗"); return; }
       const created = body?.data?.created ?? body?.created ?? 0;
       const failed = body?.data?.failed ?? body?.failed ?? 0;
       setImportDialogOpen(false);
       setImportJson("");
       await fetchTasks();
-      if (failed === 0) {
-        toast.success(`已匯入 ${created} 筆任務`);
-      } else {
-        toast.warning(`匯入完成：${created} 筆成功，${failed} 筆失敗`);
-      }
+      if (failed === 0) { toast.success(`已匯入 ${created} 筆任務`); }
+      else { toast.warning(`匯入完成：${created} 筆成功，${failed} 筆失敗`); }
     } catch {
       setImportError("網路錯誤，請稍後再試");
     } finally {
@@ -651,7 +461,7 @@ export default function KanbanPage() {
     }
   }
 
-  // ── Derived data ────────────────────────────────────────────────────────
+  // ── Derived data ──────────────────────────────────────────────────────────
 
   const tasksByStatusMap = useMemo(() => {
     const map = new Map<TaskStatus, (TaskCardData & { position?: number })[]>();
@@ -661,12 +471,8 @@ export default function KanbanPage() {
       list.push(task);
       map.set(status, list);
     }
-    // Sort each column by position
     for (const [status, list] of map) {
-      map.set(
-        status,
-        [...list].sort((a, b) => (a.position ?? 0) - (b.position ?? 0))
-      );
+      map.set(status, [...list].sort((a, b) => (a.position ?? 0) - (b.position ?? 0)));
     }
     return map;
   }, [tasks]);
@@ -678,12 +484,8 @@ export default function KanbanPage() {
       {/* Header */}
       <div className="flex flex-col sm:flex-row sm:items-center justify-between gap-2 flex-shrink-0">
         <div>
-          <h1 className="text-lg sm:text-xl font-semibold tracking-tight">
-            看板
-          </h1>
-          <p className="text-muted-foreground text-xs sm:text-sm mt-0.5">
-            共 {tasks.length} 項任務
-          </p>
+          <h1 className="text-lg sm:text-xl font-semibold tracking-tight">看板</h1>
+          <p className="text-muted-foreground text-xs sm:text-sm mt-0.5">共 {tasks.length} 項任務</p>
         </div>
         <div className="flex items-center gap-2 w-full sm:w-auto">
           {/* Multi-select toggle (Issue #1023) */}
@@ -709,7 +511,6 @@ export default function KanbanPage() {
             <FileInput className="h-3.5 w-3.5" />
             <span className="hidden sm:inline">從範本匯入</span>
           </button>
-
           <button
             onClick={handleCreateTask}
             className="flex items-center justify-center gap-1.5 text-sm font-medium px-3 py-1.5 bg-primary text-primary-foreground rounded-lg shadow-sm transition-all hover:opacity-90 flex-1 sm:flex-none"
@@ -728,10 +529,7 @@ export default function KanbanPage() {
             <strong>拖曳卡片</strong>到不同欄位即可變更任務狀態
           </span>
           <button
-            onClick={() => {
-              setShowDragHint(false);
-              localStorage.setItem("titan-kanban-onboarded", "1");
-            }}
+            onClick={() => { setShowDragHint(false); localStorage.setItem("titan-kanban-onboarded", "1"); }}
             className="ml-auto text-muted-foreground hover:text-foreground shrink-0"
           >
             <X className="h-3.5 w-3.5" />
@@ -740,317 +538,77 @@ export default function KanbanPage() {
       )}
 
       {/* Filters */}
-      <div className="flex-shrink-0 space-y-2">
-        <TaskFilters
-          filters={filters}
-          onChange={setFilters}
-          totalCount={allTasks.length}
-          filteredCount={tasks.length}
-          syncUrl
-        />
-        {/* Project filter — Issue #1176 */}
-        {projectOptions.length > 0 && (
-          <div className="flex items-center gap-2">
-            <select
-              aria-label="篩選項目"
-              value={projectFilter}
-              onChange={(e) => setProjectFilter(e.target.value)}
-              className="h-9 bg-card border border-border text-foreground text-sm rounded-lg px-3 focus:outline-none focus:border-primary focus:ring-2 focus:ring-primary/10 cursor-pointer hover:border-muted-foreground/30 transition-all shadow-sm"
-            >
-              <option value="">所有項目</option>
-              {projectOptions.map((p) => (
-                <option key={p.id} value={p.id}>
-                  {p.code} {p.name}
-                </option>
-              ))}
-            </select>
-            {projectFilter && (
-              <button
-                onClick={() => setProjectFilter("")}
-                className="text-xs text-muted-foreground hover:text-foreground"
-              >
-                <X className="h-3 w-3" />
-              </button>
-            )}
-          </div>
-        )}
-      </div>
+      <KanbanFilters
+        filters={filters}
+        allTaskCount={allTasks.length}
+        filteredTaskCount={tasks.length}
+        projectFilter={projectFilter}
+        projectOptions={projectOptions}
+        onFiltersChange={setFilters}
+        onProjectFilterChange={setProjectFilter}
+      />
 
-      {/* Bulk action bar — visible when tasks are selected in multi-select mode */}
+      {/* Bulk action bar */}
       {hasSelection && (
-        <div className="flex-shrink-0 flex flex-wrap items-center gap-2 sm:gap-3 px-3 sm:px-4 py-2.5 bg-primary/5 border border-primary/20 rounded-lg">
-          <div className="flex items-center gap-2 text-sm font-medium text-primary">
-            <CheckSquare className="h-4 w-4" />
-            已選取 {selectedIds.size} 項
-          </div>
-
-          <div className="h-4 w-px bg-border" />
-
-          {/* Batch status change */}
-          <select
-            className="text-sm border border-border rounded-md px-2 py-1 bg-background"
-            defaultValue=""
-            disabled={bulkLoading}
-            onChange={(e) => {
-              if (e.target.value) {
-                executeBulkAction({ status: e.target.value });
-                e.target.value = "";
-              }
-            }}
-            data-testid="bulk-status-select"
-          >
-            <option value="" disabled>
-              批次移動狀態
-            </option>
-            {STATUS_OPTIONS.map((o) => (
-              <option key={o.value} value={o.value}>
-                {o.label}
-              </option>
-            ))}
-          </select>
-
-          {/* Change priority */}
-          <select
-            className="text-sm border border-border rounded-md px-2 py-1 bg-background"
-            defaultValue=""
-            disabled={bulkLoading}
-            onChange={(e) => {
-              if (e.target.value) {
-                executeBulkAction({ priority: e.target.value });
-                e.target.value = "";
-              }
-            }}
-          >
-            <option value="" disabled>
-              變更優先度
-            </option>
-            {PRIORITY_OPTIONS.map((o) => (
-              <option key={o.value} value={o.value}>
-                {o.label}
-              </option>
-            ))}
-          </select>
-
-          {/* Batch assignee change */}
-          <select
-            className="text-sm border border-border rounded-md px-2 py-1 bg-background"
-            defaultValue=""
-            disabled={bulkLoading}
-            onChange={(e) => {
-              if (e.target.value) {
-                const val =
-                  e.target.value === "__unassign__" ? null : e.target.value;
-                executeBulkAction({ primaryAssigneeId: val });
-                e.target.value = "";
-              }
-            }}
-            data-testid="bulk-assignee-select"
-          >
-            <option value="" disabled>
-              批次指派
-            </option>
-            <option value="__unassign__">取消指派</option>
-            {users.map((u) => (
-              <option key={u.id} value={u.id}>
-                {u.name}
-              </option>
-            ))}
-          </select>
-
-          {/* Bulk deadline (Issue #1071) */}
-          <input
-            type="date"
-            onChange={(e) => {
-              if (e.target.value) {
-                executeBulkAction({ dueDate: new Date(e.target.value).toISOString() });
-                e.target.value = "";
-              }
-            }}
-            className="h-7 px-2 text-xs border border-border rounded-md bg-background text-foreground cursor-pointer"
-            title="批次設定截止日"
-          />
-
-          {bulkLoading && (
-            <Loader2 className="h-4 w-4 animate-spin text-muted-foreground" />
-          )}
-
-          <button
-            onClick={() => {
-              clearSelection();
-              setMultiSelectMode(false);
-            }}
-            className="ml-auto p-1 rounded hover:bg-accent text-muted-foreground hover:text-foreground transition-colors"
-            aria-label="取消選取"
-          >
-            <X className="h-4 w-4" />
-          </button>
-        </div>
+        <BulkActionBar
+          selectedCount={selectedIds.size}
+          bulkLoading={bulkLoading}
+          users={users}
+          onBulkAction={executeBulkAction}
+          onClearSelection={() => { clearSelection(); setMultiSelectMode(false); }}
+        />
       )}
 
       {/* Board */}
-      {loading ? (
-        <div className="flex-1">
-          <PageLoading message="載入看板..." />
-        </div>
-      ) : fetchError ? (
-        <div className="flex-1">
-          <PageError message={fetchError} onRetry={fetchTasks} />
-        </div>
-      ) : tasks.length === 0 && !hasActiveFilters(filters) ? (
-        <div className="flex-1">
-          <PageEmpty
-            icon={<Kanban className="h-10 w-10" />}
-            title="建立第一個任務開始工作"
-            description="把待辦事項加入看板，拖曳卡片即可更新進度"
-            action={
-              <button
-                onClick={handleCreateTask}
-                className="flex items-center gap-1.5 px-4 py-2 text-sm font-medium bg-primary text-primary-foreground rounded-lg hover:opacity-90 transition-all"
-              >
-                <Plus className="h-3.5 w-3.5" />
-                建立任務
-              </button>
-            }
-          />
-        </div>
-      ) : (
-        <div
-          className="flex-1 flex flex-col md:flex-row gap-3 overflow-y-auto md:overflow-x-auto md:overflow-y-hidden pb-4 min-h-0"
-          role="region"
-          aria-label="看板欄位"
-        >
-          {DEFAULT_COLUMNS.map(({ status, label, color }) => (
-            <div key={status} className="group/col">
-              <KanbanColumn
-                status={status}
-                label={label}
-                color={color}
-                borderClass={columnBorder[status]}
-                headerBgClass={columnHeaderBg[status]}
-                tasks={tasksByStatusMap.get(status) ?? EMPTY_TASK_ARRAY}
-                draggingId={draggingId}
-                isDragOver={dragOver === status}
-                movingTaskId={movingTask}
-                selectedIds={selectedIds}
-                hasSelection={multiSelectMode || hasSelection}
-                customName={columnNames[status]}
-                onColumnNameChange={handleColumnNameChange}
-                onDragStart={handleDragStart}
-                onDragEnd={handleDragEnd}
-                onDragOver={handleDragOver}
-                onDragLeave={handleDragLeave}
-                onDrop={handleDrop}
-                onTaskClick={setSelectedTaskId}
-                onToggleSelect={toggleSelect}
-                onKeyboardMove={handleKeyboardMove}
-                onKeyboardReorder={handleKeyboardReorder}
-              />
-            </div>
-          ))}
-        </div>
-      )}
+      <KanbanBoard
+        tasks={tasks}
+        loading={loading}
+        fetchError={fetchError}
+        filters={filters}
+        draggingId={draggingId}
+        dragOver={dragOver}
+        movingTask={movingTask}
+        selectedIds={selectedIds}
+        multiSelectMode={multiSelectMode}
+        columnNames={columnNames}
+        tasksByStatusMap={tasksByStatusMap}
+        onCreateTask={handleCreateTask}
+        onRetry={fetchTasks}
+        onDragStart={handleDragStart}
+        onDragEnd={handleDragEnd}
+        onDragOver={handleDragOver}
+        onDragLeave={handleDragLeave}
+        onDrop={handleDrop}
+        onTaskClick={setSelectedTaskId}
+        onToggleSelect={toggleSelect}
+        onKeyboardMove={handleKeyboardMove}
+        onKeyboardReorder={handleKeyboardReorder}
+        onColumnNameChange={handleColumnNameChange}
+      />
 
       {/* Task detail modal */}
       {selectedTaskId && (
         <TaskDetailModal
           taskId={selectedTaskId}
           onClose={() => setSelectedTaskId(null)}
-          onUpdated={() => {
-            setSelectedTaskId(null);
-            fetchTasks();
-          }}
+          onUpdated={() => { setSelectedTaskId(null); fetchTasks(); }}
         />
       )}
 
       {/* Template import dialog (Issue #1266) */}
       {importDialogOpen && (
-        <div
-          className="fixed inset-0 z-50 flex items-center justify-center bg-black/50 backdrop-blur-sm"
-          onClick={(e) => { if (e.target === e.currentTarget) setImportDialogOpen(false); }}
-        >
-          <div className="bg-card border border-border rounded-xl shadow-2xl w-full max-w-lg mx-4 p-6 space-y-4">
-            <div className="flex items-center justify-between">
-              <div className="flex items-center gap-2">
-                <FileInput className="h-5 w-5 text-primary" />
-                <h2 className="text-base font-semibold">從範本匯入任務</h2>
-              </div>
-              <button
-                onClick={() => setImportDialogOpen(false)}
-                className="p-1 rounded hover:bg-accent text-muted-foreground hover:text-foreground transition-colors"
-              >
-                <X className="h-4 w-4" />
-              </button>
-            </div>
-
-            <div className="text-xs text-muted-foreground space-y-1">
-              <p>貼上 JSON 範本（陣列或含 tasks 欄位的物件）。每筆任務支援欄位：</p>
-              <code className="block bg-muted/50 rounded px-2 py-1 text-[11px] leading-relaxed">
-                {`{ "title": "任務名稱", "priority": "P1", "category": "PLANNED", "status": "BACKLOG", "offsetDays": 7 }`}
-              </code>
-            </div>
-
-            <textarea
-              value={importJson}
-              onChange={(e) => { setImportJson(e.target.value); setImportError(null); }}
-              placeholder={`[\n  { "title": "任務一", "priority": "P1" },\n  { "title": "任務二", "offsetDays": 3 }\n]`}
-              rows={8}
-              className="w-full bg-muted/30 border border-border rounded-lg px-3 py-2 text-sm font-mono resize-y focus:outline-none focus:border-primary focus:ring-2 focus:ring-primary/10 transition-all"
-            />
-
-            {importError && (
-              <p className="text-xs text-red-400 flex items-center gap-1">
-                <X className="h-3 w-3 shrink-0" />
-                {importError}
-              </p>
-            )}
-
-            <div className="flex items-center justify-end gap-2 pt-1">
-              <button
-                onClick={() => setImportDialogOpen(false)}
-                className="text-sm px-4 py-1.5 rounded-lg border border-border hover:bg-accent transition-colors text-muted-foreground hover:text-foreground"
-              >
-                取消
-              </button>
-              <button
-                onClick={handleImportTemplate}
-                disabled={importLoading || !importJson.trim()}
-                className="flex items-center gap-1.5 text-sm font-medium px-4 py-1.5 bg-primary text-primary-foreground rounded-lg shadow-sm transition-all hover:opacity-90 disabled:opacity-50 disabled:cursor-not-allowed"
-              >
-                {importLoading && <Loader2 className="h-3.5 w-3.5 animate-spin" />}
-                匯入
-              </button>
-            </div>
-          </div>
-        </div>
+        <TemplateImportDialog
+          importJson={importJson}
+          importLoading={importLoading}
+          importError={importError}
+          onJsonChange={(v) => { setImportJson(v); setImportError(null); }}
+          onImport={handleImportTemplate}
+          onClose={() => setImportDialogOpen(false)}
+        />
       )}
 
       {/* Bulk operation toast (Issue #1023) */}
-      {bulkToast && (
-        <div
-          className={cn(
-            "fixed bottom-6 left-1/2 -translate-x-1/2 z-50 flex items-center gap-2 px-4 py-2.5 rounded-lg shadow-lg text-sm font-medium",
-            "animate-in fade-in slide-in-from-bottom-2 duration-200",
-            bulkToast.type === "success" &&
-              "bg-emerald-500/10 text-emerald-500 border border-emerald-500/30",
-            bulkToast.type === "error" &&
-              "bg-red-500/10 text-red-400 border border-red-500/30",
-            bulkToast.type === "partial" &&
-              "bg-amber-500/10 text-amber-500 border border-amber-500/30"
-          )}
-          role="status"
-          aria-live="polite"
-          data-testid="bulk-toast"
-        >
-          {bulkToast.type === "success" ? (
-            <Check className="h-4 w-4" />
-          ) : bulkToast.type === "error" ? (
-            <X className="h-4 w-4" />
-          ) : (
-            <CheckSquare className="h-4 w-4" />
-          )}
-          {bulkToast.message}
-        </div>
-      )}
+      {bulkToast && <BulkOperationToast toast={bulkToast} />}
     </div>
   );
 }
