+++ app/components/kpi/kpi-form-modal.tsx
@@ -1,0 +1,7 @@
+// Barrel re-export — consumers may import from either the individual
+// form files or this combined module.
+export { CreateKPIForm } from "./kpi-create-form";
+export type { CreateKPIFormProps } from "./kpi-create-form";
+
+export { EditKPIForm } from "./kpi-edit-form";
+export type { EditKPIFormProps } from "./kpi-edit-form";
